fix import cycle build error, moved score utils to its own file
