+++ x/emissions/keeper/actor_utils/util_sort.go
@@ -2,8 +2,6 @@ package actor_utils
 
 import (
 	"container/heap"
-	"context"
-	"github.com/allora-network/allora-chain/x/emissions/keeper"
 	"math/rand"
 	"sort"
 
@@ -95,82 +93,4 @@ func FindTopNByScoreDesc(n uint64, scoresByActor map[Actor]Score, randSeed Block
 	}
 
 	return topN
-}
-
-// Return low score and index among all inferences
-func GetLowScoreFromAllLossBundles(
-	ctx context.Context,
-	k *keeper.Keeper,
-	topicId TopicId,
-	lossBundles types.ReputerValueBundles,
-) (types.Score, int, error) {
-
-	lowScoreIndex := 0
-	lowScore, err := k.GetLatestReputerScore(ctx, topicId, lossBundles.ReputerValueBundles[0].ValueBundle.Reputer)
-	if err != nil {
-		return types.Score{}, lowScoreIndex, err
-	}
-	for index, extLossBundle := range lossBundles.ReputerValueBundles {
-		extScore, err := k.GetLatestReputerScore(ctx, topicId, extLossBundle.ValueBundle.Reputer)
-		if err != nil {
-			continue
-		}
-		if lowScore.Score.Lt(extScore.Score) {
-			lowScore = extScore
-			lowScoreIndex = index
-		}
-	}
-	return lowScore, lowScoreIndex, nil
-}
-
-// Return low score and index among all inferences
-func GetLowScoreFromAllInferences(
-	ctx context.Context,
-	k *keeper.Keeper,
-	topicId TopicId,
-	inferences types.Inferences,
-) (types.Score, int, error) {
-
-	lowScoreIndex := 0
-	lowScore, err := k.GetLatestInfererScore(ctx, topicId, inferences.Inferences[0].Inferer)
-	if err != nil {
-		return types.Score{}, lowScoreIndex, err
-	}
-	for index, extInference := range inferences.Inferences {
-		extScore, err := k.GetLatestInfererScore(ctx, topicId, extInference.Inferer)
-		if err != nil {
-			continue
-		}
-		if lowScore.Score.Lt(extScore.Score) {
-			lowScore = extScore
-			lowScoreIndex = index
-		}
-	}
-	return lowScore, lowScoreIndex, nil
-}
-
-// Return low score and index among all forecasts
-func GetLowScoreFromAllForecasts(
-	ctx context.Context,
-	k *keeper.Keeper,
-	topicId TopicId,
-	forecasts types.Forecasts,
-) (types.Score, int, error) {
-
-	lowScoreIndex := 0
-	lowScore, err := k.GetLatestForecasterScore(ctx, topicId, forecasts.Forecasts[0].Forecaster)
-	if err != nil {
-		return types.Score{}, lowScoreIndex, err
-	}
-	for index, extForecast := range forecasts.Forecasts {
-		extScore, err := k.GetLatestInfererScore(ctx, topicId, extForecast.Forecaster)
-		if err != nil {
-			continue
-		}
-		if lowScore.Score.Lt(extScore.Score) {
-			lowScore = extScore
-			lowScoreIndex = index
-		}
-	}
-	return lowScore, lowScoreIndex, nil
 }
